refactor: rename `sprinkles.css.ts` to `Frame.css.ts`
